teste-argo-back/requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.8
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
gunicorn==21.2.0
inflection==0.5.1
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
whitenoise==6.6.0
xlrd==2.0.1
